requirements.txt
streamlit
ultralytics
stable-baselines3
opencv-python-headless
gymnasium
numpy
scipy
pillow